+++ src/models/domain-config.ts
@@ -9,8 +9,6 @@ import {evaluateBoolean} from "../utils";
 import ApiGatewayMap = require("./api-gateway-map");
 
 class DomainConfig {
-    private readonly _stage: string | undefined;
-
     public givenDomainName: string;
     public basePath: string | undefined;
     public stage: string | undefined;
@@ -54,8 +52,18 @@ class DomainConfig {
         this.autoDomain = evaluateBoolean(config.autoDomain, false);
         this.autoDomainWaitFor = config.autoDomainWaitFor;
         this.preserveExternalPathMappings = evaluateBoolean(config.preserveExternalPathMappings, false);
+        this.basePath = this._getBasePath(config.basePath);
+        this.stage = this._getStage(config.stage);
+        this.endpointType = this._getEndpointType(config.endpointType);
+        this.apiType = this._getApiType(config.apiType);
+        this.tlsTruststoreUri = this._getTLSTruststoreUri(config.tlsTruststoreUri, this.endpointType);
+        this.tlsTruststoreVersion = config.tlsTruststoreVersion;
+        this.securityPolicy = this._getSecurityPolicy(config.securityPolicy);
+        this.route53Params = this._getRoute53Params(config.route53Params, this.endpointType);
+        this.splitHorizonDns = !this.hostedZoneId && !this.hostedZonePrivate && evaluateBoolean(config.splitHorizonDns, false);
+    }
 
-        let basePath = config.basePath;
+    private _getBasePath(basePath: string | undefined) {
         if (!basePath || basePath.trim() === "") {
             basePath = Globals.defaultBasePath;
         }
@@ -65,79 +73,82 @@ class DomainConfig {
                 "https://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-known-issues.html"
             );
         }
-        this.basePath = basePath;
+        return basePath;
+    }
 
-        this._stage = config.stage;
-        this.stage = config.stage || Globals.options.stage || Globals.serverless.service.provider.stage;
+    private _getStage(stage: string | undefined) {
+        return stage || Globals.options.stage || Globals.serverless.service.provider.stage;
+    }
 
-        const endpointTypeWithDefault = config.endpointType || Globals.endpointTypes.edge;
+    private _getEndpointType(endpointType: string | undefined) {
+        const endpointTypeWithDefault = endpointType || Globals.endpointTypes.edge;
         const endpointTypeToUse = Globals.endpointTypes[endpointTypeWithDefault.toLowerCase()];
         if (!endpointTypeToUse) {
             throw new Error(`${endpointTypeWithDefault} is not supported endpointType, use EDGE or REGIONAL.`);
         }
-        this.endpointType = endpointTypeToUse;
+        return endpointTypeToUse;
+    }
 
-        const apiTypeWithDefault = config.apiType || Globals.apiTypes.rest;
+    private _getApiType(apiType: string | undefined) {
+        const apiTypeWithDefault = apiType || Globals.apiTypes.rest;
         const apiTypeToUse = Globals.apiTypes[apiTypeWithDefault.toLowerCase()];
         if (!apiTypeToUse) {
             throw new Error(`${apiTypeWithDefault} is not supported api type, use REST, HTTP or WEBSOCKET.`);
         }
-        this.apiType = apiTypeToUse;
+        return apiTypeToUse;
+    }
 
-        const isEdgeType = this.endpointType === Globals.endpointTypes.edge;
-        if (isEdgeType && config.tlsTruststoreUri) {
-            throw new Error(`${this.endpointType} APIs do not support mutual TLS, remove tlsTruststoreUri or change to a regional API.`);
-        }
-        if (config.tlsTruststoreUri) {
-            this.validateS3Uri(config.tlsTruststoreUri);
-        }
-        this.tlsTruststoreUri = config.tlsTruststoreUri;
-        this.tlsTruststoreVersion = config.tlsTruststoreVersion;
+    private _getTLSTruststoreUri(tlsTruststoreUri: string | undefined, endpointType: string) {
+        if (tlsTruststoreUri) {
+            if (endpointType === Globals.endpointTypes.edge) {
+                throw new Error(
+                    endpointType + " APIs do not support mutual TLS, " +
+                    "remove tlsTruststoreUri or change to a regional API."
+                );
+            }
 
-        const securityPolicyDefault = config.securityPolicy || Globals.tlsVersions.tls_1_2;
+            const {protocol, pathname} = new URL(tlsTruststoreUri);
+
+            if (protocol !== "s3:" && !pathname.substring(1).includes("/")) {
+                throw new Error(
+                    `${tlsTruststoreUri} is not a valid s3 uri, try something like s3://bucket-name/key-name.`
+                );
+            }
+        }
+
+        return tlsTruststoreUri;
+    }
+
+    private _getSecurityPolicy(securityPolicy: string | undefined) {
+        const securityPolicyDefault = securityPolicy || Globals.tlsVersions.tls_1_2;
         const tlsVersionToUse = Globals.tlsVersions[securityPolicyDefault.toLowerCase()];
         if (!tlsVersionToUse) {
             throw new Error(`${securityPolicyDefault} is not a supported securityPolicy, use tls_1_0 or tls_1_2.`);
         }
-        this.securityPolicy = tlsVersionToUse;
 
-        const defaultRoutingPolicy = Globals.routingPolicies.simple;
-        const routingPolicy = config.route53Params?.routingPolicy?.toLowerCase() ?? defaultRoutingPolicy;
+        return tlsVersionToUse;
+    }
+
+    private _getRoute53Params(route53Params: Route53Params | undefined, endpointType: string) {
+        const routingPolicy = route53Params?.routingPolicy?.toLowerCase() ?? Globals.routingPolicies.simple;
         const routingPolicyToUse = Globals.routingPolicies[routingPolicy];
         if (!routingPolicyToUse) {
             throw new Error(`${routingPolicy} is not a supported routing policy, use simple, latency, or weighted.`);
         }
 
-        if (routingPolicyToUse !== defaultRoutingPolicy && endpointTypeToUse === Globals.endpointTypes.edge) {
+        if (routingPolicyToUse !== Globals.routingPolicies.simple && endpointType === Globals.endpointTypes.edge) {
             throw new Error(
                 `${routingPolicy} routing is not intended to be used with edge endpoints. ` +
                 "Use a regional endpoint instead."
             );
         }
 
-        this.route53Params = {
+        return {
             routingPolicy: routingPolicyToUse,
-            setIdentifier: config.route53Params?.setIdentifier,
-            weight: config.route53Params?.weight ?? 200,
-            healthCheckId: config.route53Params?.healthCheckId
+            setIdentifier: route53Params?.setIdentifier,
+            weight: route53Params?.weight ?? 200,
+            healthCheckId: route53Params?.healthCheckId
         };
-
-        this.splitHorizonDns = !this.hostedZoneId && !this.hostedZonePrivate && evaluateBoolean(config.splitHorizonDns, false);
-    }
-
-    /**
-     * Return the stage provided in the custom config
-     */
-    public getConfigStage(): string | undefined {
-        return this._stage;
-    }
-
-    private validateS3Uri(uri: string): void {
-        const {protocol, pathname} = new URL(uri);
-
-        if (protocol !== "s3:" && !pathname.substring(1).includes("/")) {
-            throw new Error(`${uri} is not a valid s3 uri, try something like s3://bucket-name/key-name.`);
-        }
     }
 }